+++ Libraries/Cordio/platform/targets/maxim/max32655/sources/pal_flash.c
@@ -121,7 +121,7 @@ PalFlashState_t PalNvmGetState(void)
 /*************************************************************************************************/
 uint32_t PalNvmGetTotalSize(void)
 {
-  return &__pal_nvm_db_end__ - &__pal_nvm_db_start__;
+  return (uint32_t)&__pal_nvm_db_end__ - (uint32_t)&__pal_nvm_db_start__;
 }
 
 /*************************************************************************************************/
